general push+ note for PS5
A note on PS5: I'm going to be out of town on non-WashU business for a five day span (I've talked to Jacob about this) and have basically timed myself out of actually working through PS5. I'm basically taking the zero and drop on this one, but if you look for my submission and find it blank, that is why.